Fix checkpoint typing
