+++ setup.cfg
@@ -41,7 +41,7 @@ install_requires =
     numba
     pandas
     pyarrow >= 14.0.0
-    pydantic
+    pydantic < 2.0.0
     pyyaml >= 5.1
     quivr @ git+https://github.com/moeyensj/quivr@concatenate-empty-attributes
     ray[default]
